useHandlePastedText hook
